web_scraper.py
"""
Web scraping module for social media platforms using Gemini Vision
"""

import json
from typing import Dict, List, Optional
from config import SCRAPER_USER_AGENT, SCRAPER_TIMEOUT, DEFAULT_TIME_RANGE_DAYS


class WebScraperError(Exception):
    """Custom exception for web scraping errors"""
    pass


class AgenticScraper:
    """
    Gemini-powered web scraper for creator profiles

    Uses Gemini's multimodal capabilities to:
    1. Screenshot profile pages
    2. Extract structured data from screenshots
    3. Navigate paginated content
    """

    def __init__(self, gemini_api_key: str):
        """
        Initialize agentic scraper

        Args:
            gemini_api_key: Gemini API key for vision analysis
        """
        self.api_key = gemini_api_key

        # Try importing required libraries
        try:
            import google.generativeai as genai
            genai.configure(api_key=gemini_api_key)
            self.genai = genai
            self.vision_available = True
        except ImportError:
            print("Gemini API not available")
            self.vision_available = False

    def scrape_creator_profile(self, url: str, platform: str) -> Dict:
        """
        Scrape creator profile using Gemini Vision

        Args:
            url: Profile URL to scrape
            platform: Platform name (instagram, tiktok, twitch)

        Returns:
            Dict with profile statistics
        """
        raise NotImplementedError(
            "Web scraping with Gemini Vision will be implemented in a future phase. "
            "For now, use YouTube API via platform_clients.py"
        )

    def extract_recent_posts(self, url: str, days: int = DEFAULT_TIME_RANGE_DAYS,
                           max_posts: int = 50) -> List[Dict]:
        """
        Extract recent posts from a profile

        Args:
            url: Profile URL
            days: Number of days to look back
            max_posts: Maximum posts to extract

        Returns:
            List of post dictionaries
        """
        raise NotImplementedError(
            "Post extraction with Gemini Vision will be implemented in a future phase"
        )

    def discover_alternate_accounts(self, creator_name: str, known_platforms: List[str] = None) -> List[Dict]:
        """
        Discover alternate social media accounts for a creator

        Uses Gemini to search the web and identify likely account matches

        Args:
            creator_name: Name of the creator
            known_platforms: List of platforms already known

        Returns:
            List of discovered account dictionaries
        """
        raise NotImplementedError(
            "Account discovery will be implemented in a future phase. "
            "For now, users can manually add alternate accounts."
        )


# Helper function to detect platform from URL
def detect_platform_from_url(url: str) -> Optional[str]:
    """
    Detect platform from URL

    Args:
        url: Social media URL

    Returns:
        Platform name (youtube, instagram, tiktok, twitch) or None
    """
    url_lower = url.lower()

    if 'youtube.com' in url_lower or 'youtu.be' in url_lower:
        return 'youtube'
    elif 'instagram.com' in url_lower:
        return 'instagram'
    elif 'tiktok.com' in url_lower:
        return 'tiktok'
    elif 'twitch.tv' in url_lower:
        return 'twitch'
    else:
        return None


def extract_handle_from_url(url: str, platform: str) -> str:
    """
    Extract username/handle from social media URL

    Args:
        url: Social media URL
        platform: Platform name

    Returns:
        Username/handle
    """
    url = url.rstrip('/')

    if platform == 'youtube':
        if '/@' in url:
            return url.split('/@')[-1].split('?')[0]
        elif '/channel/' in url:
            return url.split('/channel/')[-1].split('?')[0]
        elif '/c/' in url:
            return url.split('/c/')[-1].split('?')[0]
        elif '/user/' in url:
            return url.split('/user/')[-1].split('?')[0]
    elif platform == 'instagram':
        # https://instagram.com/username/
        parts = url.split('instagram.com/')
        if len(parts) > 1:
            return parts[1].split('/')[0].split('?')[0]
    elif platform == 'tiktok':
        # https://tiktok.com/@username
        if '/@' in url:
            return url.split('/@')[-1].split('?')[0]
    elif platform == 'twitch':
        # https://twitch.tv/username
        parts = url.split('twitch.tv/')
        if len(parts) > 1:
            return parts[1].split('/')[0].split('?')[0]

    # Fallback: return last part of URL
    return url.split('/')[-1].split('?')[0]
